==> 1_Mining.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

url = "https://raw.githubusercontent.com/fahmiagungmaulana/data1/main/train.csv"
train = pd.read_csv(url)

#Mining Data Mulai

train['id_titik_mulai'].nunique()

mulai = train.id_titik_mulai.unique()

import requests
from bs4 import BeautifulSoup
import pandas as pd

def scrape_lat_long(node_id):
    url = f"https://www.openstreetmap.org/?node={node_id}"
    response = requests.get(url)

    if response.status_code == 200:
        soup = BeautifulSoup(response.text, 'html.parser')

        latitude_elem = soup.find(class_='latitude')
        longitude_elem = soup.find(class_='longitude')

        if latitude_elem and longitude_elem:
            latitude = latitude_elem.get_text()
            longitude = longitude_elem.get_text()

            return latitude, longitude
        else:
            print(f"Latitude or longitude data not found for node {node_id}.")
            return None, None
    else:
        print(f"Failed to retrieve data for node {node_id}. Status code: {response.status_code}")
        return None, None

data = []
for node_id in mulai:
    latitude, longitude = scrape_lat_long(node_id)
    if latitude and longitude:
        data.append({'node_id': node_id, 'latitude': latitude, 'longitude': longitude})


#ambil versi sebelumnya
data.append({'node_id': 5669475346, 'latitude': 51.5885105, 'longitude': -0.1998135})
data.append({'node_id': 227758, 'latitude': 51.4597463, 'longitude': -0.1291217})
data.append({'node_id': 227771, 'latitude': 51.4502291, 'longitude': -0.1225601})
data.append({'node_id': 1663461749, 'latitude': 51.6379750, 'longitude': -0.1790446})
data.append({'node_id': 1701556095, 'latitude': 51.6345474, 'longitude': -0.1761378})
data.append({'node_id': 2375143, 'latitude': 51.5324067, 'longitude': -0.2335842})
data.append({'node_id': 6205314505, 'latitude': 51.4633403, 'longitude': -0.2437799})

train_mulai = pd.DataFrame(data)
train_mulai.shape

train_mulai.head()

train_mulai.isnull().sum()

train_mulai.to_csv('train_mulai.csv', index=False)

#Mining data akhir

train.id_titik_akhir.nunique()

akhir = train.id_titik_akhir.unique()

import requests
from bs4 import BeautifulSoup
import pandas as pd

def scrape_lat_long(node_id):
    url = f"https://www.openstreetmap.org/?node={node_id}"
    response = requests.get(url)

    if response.status_code == 200:
        soup = BeautifulSoup(response.text, 'html.parser')

        latitude_elem = soup.find(class_='latitude')
        longitude_elem = soup.find(class_='longitude')

        if latitude_elem and longitude_elem:
            latitude = latitude_elem.get_text()
            longitude = longitude_elem.get_text()

            return latitude, longitude
        else:
            print(f"Latitude or longitude data not found for node {node_id}.")
            return None, None
    else:
        print(f"Failed to retrieve data for node {node_id}. Status code: {response.status_code}")
        return None, None

data_akhir = []
for node_id in akhir:
    latitude, longitude = scrape_lat_long(node_id)
    if latitude and longitude:
        data_akhir.append({'node_id': node_id, 'latitude': latitude, 'longitude': longitude})


#ambil versi sebelumnya
data_akhir.append({'node_id': 5669475346, 'latitude': 51.5885105, 'longitude': -0.1998135})
data_akhir.append({'node_id': 1701556095, 'latitude': 51.6345474, 'longitude': -0.1761378})
data_akhir.append({'node_id': 2375143, 'latitude': 51.5324067, 'longitude': -0.2335842})
data_akhir.append({'node_id': 6205314505, 'latitude': 51.4633403, 'longitude': -0.2437799})
data_akhir.append({'node_id': 227758, 'latitude': 51.4597463, 'longitude': -0.1291217})
data_akhir.append({'node_id': 227771, 'latitude': 51.4502291, 'longitude': -0.1225601})
data_akhir.append({'node_id': 1663461749, 'latitude': 51.6379750, 'longitude': -0.1790446})

train_akhir = pd.DataFrame(data_akhir)
train_akhir.shape

train_akhir.head()

train_akhir.isnull().sum()

train_akhir.to_csv('train_akhir.csv', index=False)
