==> 2_Mining Jalan.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

url = "https://raw.githubusercontent.com/fahmiagungmaulana/data1/main/train.csv"
train = pd.read_csv(url)

train.head()

#Mining Jalan

train.id_jalan.nunique()

train.id_jalan.unique()

data = {
    'id': [691007296, 47010584, 22932408, 142479648, 8504977, 50841260, 4060819, 24704286, 4061596, 4004732, 3688059, 66924592, 3691841, 97453767, 4898588, 26785612, 182210371, 150783571, 40044423, 1210],
    'maxspeed_mph': [30, 30, 20, 20, 30, 30, 20, 30, 20, 30, 30, 30, 30, 20, 20, 20, 30, 30, 30, 30]
}

df_jalan = pd.DataFrame(data)

df_jalan['maxspeed_mph'] = df_jalan['maxspeed_mph'].astype(str)

df_jalan.info()

df_jalan.to_csv("df_jalan.csv", index=False)
