==> 3_Hitung Jarak.ipynb <==
import pandas as pd
import numpy as np

url = "https://raw.githubusercontent.com/fahmiagungmaulana/data1/main/train.csv"
train = pd.read_csv(url)

train.head()

train['kelompok'] = train.apply(lambda row: '_'.join(sorted([str(row['id_titik_mulai']), str(row['id_titik_akhir'])])), axis=1)

train.kelompok.nunique()

data_kel = train.kelompok.unique()

# Inisialisasi list untuk menyimpan angka bagian kiri dan kanan
angka_kiri = []
angka_kanan = []

# Iterasi melalui setiap kelompok dan memisahkan angka
for kelompok in data_kel:
    angka = kelompok.split("_")
    angka_kiri.append(int(angka[0]))
    angka_kanan.append(int(angka[1]))

# Membuat dataframe baru
data = {'Angka_Kiri': angka_kiri, 'Angka_Kanan': angka_kanan}
df = pd.DataFrame(data)


train_mulai = pd.read_csv("train_mulai.csv")

# Merge dataframe df dengan train_mulai berdasarkan angka_kiri dan node_id
df_mulai = df.merge(train_mulai, left_on='Angka_Kiri', right_on='node_id', how='left')

# Menghapus kolom node_id agar tidak ada duplikasi
df_mulai = df_mulai.drop(columns='node_id')

df_mulai = df_mulai.rename(columns={'latitude':'lat1','longitude':'long1'})

train_akhir = pd.read_csv("train_akhir.csv")

# Merge dataframe df dengan train_mulai berdasarkan angka_kiri dan node_id
df_full = df_mulai.merge(train_akhir, left_on='Angka_Kanan', right_on='node_id', how='left')

# Menghapus kolom node_id agar tidak ada duplikasi
df_full = df_full.drop(columns='node_id')

df_full = df_full.rename(columns={'latitude':'lat2','longitude':'long2'})

from geopy.distance import distance
# Fungsi untuk menghitung jarak antara dua koordinat
def hitung_jarak(row):
    koordinat1 = (row['lat1'], row['long1'])
    koordinat2 = (row['lat2'], row['long2'])
    return distance(koordinat1, koordinat2).kilometers

# Menambahkan kolom jarak ke DataFrame
df_full['jarak_km'] = df_full.apply(hitung_jarak, axis=1)

df_full.head()

df_full.to_csv("Jarak.csv", index=False)
