==> 4_Gabungan1.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

url1 = "https://raw.githubusercontent.com/fahmiagungmaulana/data1/main/train.csv"
train = pd.read_csv(url1)

url2 = "https://raw.githubusercontent.com/fahmiagungmaulana/data1/main/test.csv"
test = pd.read_csv(url2)

#Train

df_jalan = pd.read_csv('df_jalan.csv')
jarak = pd.read_csv('Jarak.csv')

train.shape

train['kelompok'] = train.apply(lambda row: '_'.join(sorted([str(row['id_titik_mulai']), str(row['id_titik_akhir'])])), axis=1)

data_kel = train.kelompok.unique()
# Inisialisasi list untuk menyimpan angka bagian kiri dan kanan
angka_kiri = []
angka_kanan = []

# Iterasi melalui setiap kelompok dan memisahkan angka
for kelompok in data_kel:
    angka = kelompok.split("_")
    angka_kiri.append(int(angka[0]))
    angka_kanan.append(int(angka[1]))

# Membuat dataframe baru
data = {'Kelompok': data_kel,'Angka_Kiri': angka_kiri, 'Angka_Kanan': angka_kanan}
df = pd.DataFrame(data)

# Gabungkan data berdasarkan kolom id_titik_mulai dan node_id
train_combined = pd.merge(train, df, left_on='kelompok', right_on='Kelompok')
# Hapus kolom node_id karena sudah tidak diperlukan setelah penggabungan
train_combined = train_combined.drop(columns=['Kelompok'])
# train_combined = train_combined.drop(columns=['id_titik_mulai', 'id_titik_akhir'])
train_combined = train_combined.rename(columns={'Angka_Kiri':'titik1','Angka_Kanan':'titik2'})

train_combined2 = pd.merge(train_combined, df_jalan, left_on='id_jalan', right_on='id')
# Hapus kolom node_id karena sudah tidak diperlukan setelah penggabungan
train_combined2 = train_combined2.drop(columns=['id'])

jarak['Angka_Kiri'] = jarak['Angka_Kiri'].astype(str)
jarak['Angka_Kanan'] = jarak['Angka_Kanan'].astype(str)

jarak.head()

jarak['Kelompok'] = jarak.apply(lambda row: '_'.join(sorted([str(row['Angka_Kiri']), str(row['Angka_Kanan'])])), axis=1)

jarak = jarak.drop(columns=['Angka_Kiri','Angka_Kanan'])

train_combined2

# Gabungkan data berdasarkan kolom id_titik_mulai dan node_id
train_combined3 = pd.merge(train_combined2, jarak, left_on='kelompok', right_on='Kelompok')
# Hapus kolom node_id karena sudah tidak diperlukan setelah penggabungan
train_combined3 = train_combined3.drop(columns=['Kelompok'])

train_combined3.head()

train_combined3.shape

train_combined3.to_csv("train_combined.csv", index=False)

#Test

test.shape

test['kelompok'] = test.apply(lambda row: '_'.join(sorted([str(row['id_titik_mulai']), str(row['id_titik_akhir'])])), axis=1)

data_kel_test = test.kelompok.unique()
# Inisialisasi list untuk menyimpan angka bagian kiri dan kanan
angka_kiri_test = []
angka_kanan_test = []

# Iterasi melalui setiap kelompok dan memisahkan angka
for kelompok in data_kel_test:
    angka_test = kelompok.split("_")
    angka_kiri_test.append(int(angka_test[0]))
    angka_kanan_test.append(int(angka_test[1]))

# Membuat dataframe baru
data_test = {'Kelompok': data_kel_test,'Angka_Kiri': angka_kiri_test, 'Angka_Kanan': angka_kanan_test}
df_test = pd.DataFrame(data_test)

# Gabungkan data berdasarkan kolom id_titik_mulai dan node_id
test_combined = pd.merge(test, df_test, left_on='kelompok', right_on='Kelompok')
# Hapus kolom node_id karena sudah tidak diperlukan setelah penggabungan
test_combined = test_combined.drop(columns=['Kelompok'])
# train_combined = train_combined.drop(columns=['id_titik_mulai', 'id_titik_akhir'])
test_combined = test_combined.rename(columns={'Angka_Kiri':'titik1','Angka_Kanan':'titik2'})

test_combined2 = pd.merge(test_combined, df_jalan, left_on='id_jalan', right_on='id')
test_combined2 = test_combined2.drop(columns=['id_y'])

jarak = pd.read_csv('Jarak.csv')

jarak['Angka_Kiri'] = jarak['Angka_Kiri'].astype(str)
jarak['Angka_Kanan'] = jarak['Angka_Kanan'].astype(str)
jarak['Kelompok'] = jarak.apply(lambda row: '_'.join(sorted([str(row['Angka_Kiri']), str(row['Angka_Kanan'])])), axis=1)
jarak = jarak.drop(columns=['Angka_Kiri','Angka_Kanan'])

# Gabungkan data berdasarkan kolom id_titik_mulai dan node_id
test_combined3 = pd.merge(test_combined2, jarak, left_on='kelompok', right_on='Kelompok')
# Hapus kolom node_id karena sudah tidak diperlukan setelah penggabungan
test_combined3 = test_combined3.drop(columns=['Kelompok'])

test_combined3 = test_combined3.rename(columns={'id_x':'id'})

test_combined3 = test_combined3.sort_values(by='id')

test_combined3.shape

test_combined3.head()

test_combined3.to_csv("test_combined.csv", index=False)
